+++ src/utils/HADictionary.cpp
@@ -19,6 +19,7 @@ const char HANameProperty[] PROGMEM = {"name"};
 const char HAUniqueIdProperty[] PROGMEM = {"uniq_id"};
 const char HADeviceProperty[] PROGMEM = {"dev"};
 const char HADeviceClassProperty[] PROGMEM = {"dev_cla"};
+const char HAIconProperty[] PROGMEM = {"ic"};
 
 // topics
 const char HAConfigTopic[] PROGMEM = {"config"};